Inlined `lighttpd.rs` into `main.rs` in the `lighttpd` test crate.
This allows `c2rust-analyze` to easily run on everything,
as it does not yet support compiling whole crates
(a `cargo` wrapper, not the current `rustc` wrapper).
